+++ tests/13_Recursive.test.ts
@@ -22,7 +22,7 @@ describe("Recursive", () => {
     const foo = new Foo();
     const logMsg = mainLogger.info("circular", foo);
     expect(logMsg?.["0"]).toBe("circular");
-    expect((logMsg?.["1"] as any)["circular"]).toEqual(logMsg?.["1"]);
+    expect((logMsg?.["1"] as Record<string, unknown>)["circular"]).toEqual(logMsg?.["1"]);
   });
 
   test("json", (): void => {
@@ -41,7 +41,7 @@ describe("Recursive", () => {
     const foo = new Foo();
     const logMsg = mainLogger.info("circular", foo);
     expect(logMsg?.["0"]).toBe("circular");
-    expect((logMsg?.["1"] as any)["circular"]).toEqual(logMsg?.["1"]);
+    expect((logMsg?.["1"] as Record<string, unknown>)["circular"]).toEqual(logMsg?.["1"]);
   });
 
   test("pretty", (): void => {
@@ -60,7 +60,7 @@ describe("Recursive", () => {
     const foo = new Foo();
     const logMsg = mainLogger.info("circular", foo);
     expect(logMsg?.["0"]).toBe("circular");
-    expect((logMsg?.["1"] as any)["circular"]).toEqual(logMsg?.["1"]);
+    expect((logMsg?.["1"] as Record<string, unknown>)["circular"]).toEqual(logMsg?.["1"]);
   });
 
   test("pretty recursive LogObj function", (): void => {